backend/requirements.txt
Django>=5.0.1
djangorestframework
djangorestframework-authtoken
django-cors-headers
stripe
Pillow